# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AIALRA Model Router API
  version: 0.1.0
  description: Private Responses-compatible and native Jobs API for quota-aware model routing
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://router.tailnet.example
    description: Private Tailnet endpoint
security:
  - bearerAuth: []
paths:
...
  /api/v1/routes/preview:
    post:
      operationId: previewRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task]
              properties:
                task:
                  $ref: "#/components/schemas/TaskContract"
      responses:
        "201":
          description: Deterministic route decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDecision"
...
  /api/v1/audit:
    get:
      operationId: listAuditEvents
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Administrative audit events
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEvent"
  /api/v1/retention/receipts:
    get:
      operationId: listDeletionReceipts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Payload deletion receipts and metadata deletion deadlines
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DeletionReceipt"
...
  /auth/passkey/register/options:
    post:
      operationId: beginPasskeyRegistration
      security: []
      parameters:
        - $ref: "#/components/parameters/BootstrapToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                displayName:
                  type: string
      responses:
        "201":
          description: WebAuthn registration challenge and browser options
  /auth/passkey/register/verify:
    post:
      operationId: verifyPasskeyRegistration
      security: []
      parameters:
        - $ref: "#/components/parameters/BootstrapToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, response]
              properties:
                challengeId:
                  type: string
                  format: uuid
                response:
                  type: object
      responses:
        "201":
          description: Registration result and ten one-time recovery codes
  /auth/passkey/login/options:
    post:
      operationId: beginPasskeyLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "201":
          description: WebAuthn authentication challenge and browser options
  /auth/passkey/login/verify:
    post:
      operationId: verifyPasskeyLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, response]
              properties:
                challengeId:
                  type: string
                  format: uuid
                response:
                  type: object
      responses:
        "201":
          description: Authentication result with an HttpOnly session cookie
  /auth/recovery/login:
    post:
      operationId: loginWithRecoveryCode
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
      responses:
        "201":
          description: Authentication result after consuming one recovery code
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: AIALRA_API_KEY
  parameters:
...
    BootstrapToken:
      name: X-Bootstrap-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
...
  schemas:
...
    AuditEvent:
      type: object
      required: [id, actorId, action, resourceType, resourceId, metadata, createdAt]
      properties:
        id:
          type: string
          format: uuid
        actorId:
          type: string
        action:
          type: string
        resourceType:
          type: string
        resourceId:
          type: [string, "null"]
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    DeletionReceipt:
      type: object
      required: [id, jobId, payloadDeletedAt, metadataDeleteAfter]
      properties:
        id:
          type: string
          format: uuid
        jobId:
          type: string
          format: uuid
        payloadDeletedAt:
          type: string
          format: date-time
        metadataDeleteAfter:
          type: string
          format: date-time
    PermissionProfile:
      type: object
      additionalProperties: false
      properties:
        filesystem:
          type: string
          enum: [none, read, write]
          default: read
        network:
          type: string
          enum: [none, allowlist]
          default: none
        allowedHosts:
          type: array
          maxItems: 32
          items:
            type: string
        requireApprovalForWrites:
          type: boolean
          default: true
        requireApprovalForExternalActions:
          type: boolean
          default: true
    TaskContract:
      type: object
      required: [objective]
      additionalProperties: false
      properties:
        objective:
          type: string
          minLength: 1
          maxLength: 100000
        taskKind:
          type: string
          enum: [bounded, coding, review, planning, batch, general]
          default: general
        requiredContext:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        expectedOutput:
          type: string
        validation:
          type: object
          properties:
            responseSchema:
              type: object
              additionalProperties: true
            acceptanceTests:
              type: array
              items:
                type: string
        dataClassification:
          type: string
          enum: [public, internal, confidential, restricted]
          default: internal
        permissions:
          $ref: "#/components/schemas/PermissionProfile"
        deadlineMs:
          type: integer
          minimum: 1000
          maximum: 3600000
          default: 120000
        budget:
          type: object
          properties:
            maxCodexCredits:
              type: number
              exclusiveMinimum: 0
            maxOutputTokens:
              type: integer
              default: 8192
            maxAttempts:
              type: integer
              minimum: 1
              maximum: 2
              default: 2
        sessionKey:
          type: string
        model:
          type: string
          enum: [auto, luna, terra, sol]
          default: auto
        effort:
          type: string
          enum: [minimal, low, medium, high, xhigh]
          default: medium
        replayable:
          type: boolean
          default: true
        ambiguity:
          type: integer
          minimum: 0
          maximum: 4
        risk:
          type: integer
          minimum: 0
          maximum: 4
...
    RouteDecision:
      type: object
      required: [provider, model, effort, policyVersion, reasonCode, sticky]
      properties:
        provider:
          type: string
          const: codex
        model:
          type: string
        effort:
          type: string
        policyVersion:
          type: string
        reasonCode:
          type: string
        sticky:
          const: true
...